data-quality-check-app/backend/requirements.txt
flask
flask-cors
pandas
openai>=1.0.0
pyyaml
